requirements.txt
aiopg==1.4.0
discord.py==2.5.2
riotwatcher==3.3.1
